Refactor: Infinite Scroll 컴포넌트 분리
